Add drop rules for FORWARD actions for all physical interfaces
In addition to blocking FORWARD actions on breth0, we also want to block
forwarding on all physical interfaces.

If we do so, the OVNKubernetes stack can run in 3 distinct modes:
* Global: No restrictions to IP forwarding. The OS must enable
  forwarding globally via sysctl. OVNKubernetes must run without
  CLI option --disable-forwarding.
* Restricted: Disable IP forwarding on the entire system, except on
  breth0 and the mp0 interface. Block unwanted forwarding with iptables
  on breth0 and on mp0. For this to work, the OS must disable
  forwarding globally via sysctl. OVNKubernetes must run with CLI
  option --disable-forwarding.
* RestrictedPhysical: Block unwanted forwarding with iptables on
  breth0, the mp0 interface and on all physical interfaces. Allow other
  forwarding. For this to work, the OS must enable forwarding globally
  via sysctl. OVNKubernetes must run with CLI option
  --disable-forwarding.

It should be irrelevant if we add additional iptables DROP rules to the
physical interfaces in Restricted mode (forwarding is already off). It
should be irrelevant if we explicitly enable forwarding on breth0 and
mp0 in RestrictedPhysical mode (it will already be on globally).
Therefore, OVNKubernetes can live with a single command line flag.

Signed-off-by: Andreas Karis <ak.karis@gmail.com>